Inject list input variables from query to field
In DocumentExecuter.GetVariableValue the variables type's name was
retrieved. This stripped list and non-nullable information and made the
IsValidValue check fail.

This change instead gets the type via the GraphTypeFromType method. The
test executes_with_injected_input_variables is extended to cover the
scenario. The properly_parses_multiple_values_to_list does not demonstrate
a bug but is there for completeness.

Additionally GetArgument on context did not support IEnumerable as generic
parameter. This is fixed in ObjectExtensions and tested in
can_get_list_argument.

Signed-off-by: Lasse Nisted <nistednistednisted@gmail.com>